+++ web/packages/teleport/src/Account/index.ts
@@ -16,5 +16,5 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-import Account from './AccountNew';
+import Account from './Account';
 export default Account;
